django/requirements.txt
django
django-cors-headers
psycopg2-binary
django-oauth-toolkit
djangorestframework
django-filter
requests
pyjwt